requirements.txt
allure-pytest==2.13.0
pytest==7.2.2
pytest-xdist==3.2.0
pytest-metadata==2.0.4
Appium-Python-Client==2.8.1